refactor(patina): split no_use_computed_property_like_method tests into submodule
